+++ lib/src/features/training_log/domain/repositories/workout_set_repository.dart
@@ -1,6 +1,7 @@
 import '../entities/workout_set.dart';
 import '../value_objects/exercise_ref.dart';
 import '../value_objects/stable_ids.dart';
+import '../value_objects/workout_set_timeline.dart';
 
 abstract interface class WorkoutSetRepository {
   Future<void> save(WorkoutSet set);
@@ -8,6 +9,10 @@ abstract interface class WorkoutSetRepository {
   Future<WorkoutSet?> findById(WorkoutSetId id);
 
   Future<List<WorkoutSet>> historyForExercise(ExerciseRef exerciseRef);
+
+  Future<WorkoutSetTimelinePage> timelineForExercise(
+    WorkoutSetTimelineQuery query,
+  );
 
   Future<List<WorkoutSet>> setsForWorkoutSession(
     WorkoutSessionId workoutSessionId,